+++ apps/ui/src/components/views/project-settings-view/project-settings-view.tsx
@@ -5,8 +5,7 @@ import { Button } from '@/components/ui/button';
 import { ProjectIdentitySection } from './project-identity-section';
 import { ProjectThemeSection } from './project-theme-section';
 import { WorktreePreferencesSection } from './worktree-preferences-section';
-import { TestingSection } from './testing-section';
-import { DevServerSection } from './dev-server-section';
+import { CommandsSection } from './commands-section';
 import { ProjectModelsSection } from './project-models-section';
 import { DataManagementSection } from './data-management-section';
 import { DangerZoneSection } from '../settings-view/danger-zone/danger-zone-section';
@@ -88,10 +87,8 @@ export function ProjectSettingsView() {
         return <ProjectThemeSection project={currentProject} />;
       case 'worktrees':
         return <WorktreePreferencesSection project={currentProject} />;
-      case 'testing':
-        return <TestingSection project={currentProject} />;
-      case 'devServer':
-        return <DevServerSection project={currentProject} />;
+      case 'commands':
+        return <CommandsSection project={currentProject} />;
       case 'claude':
         return <ProjectModelsSection project={currentProject} />;
       case 'data':
